Create Netcat-Reverseshell.ino
This script allows digispark to disable windows defender. Allows you to download netcat from http web with powershell and run netcat by opening the appropriate port. You can also change the order of the "/t" tabs in the script for your situation. I'm still finding a way to bypass the firewall, which prevents netcat from running instantly. to connect to the machine just type "nc victim ip port". To find the ip just add the script "Grab-WindowsSysinfo.ino" which is in that same repository.